+++ tests/utilities.js
@@ -61,3 +61,8 @@ export let wrap_front = function(obj, fn_name, fn, is_setter=false) {
 		active: true
 	});
 };
+
+export let unwrap_all_from_obj = function(obj, fn_name, is_setter=false) {
+	const wrapper = libWrapper._create_wrapper_from_object(obj, fn_name);
+	wrapper.get_fn_data(is_setter).splice(0);
+}